update documentation and cmake

+++ headers/layout.h
@@ -18,7 +18,7 @@ public:
 
 private:
   void fruchtermanReingold(const std::shared_ptr<Graph> &graph, int iterations,
-                           float k, int width, int height, int x, int y);
+                           float k, double width, double height, int x, int y);
   void layoutBruteForce(const std::shared_ptr<Graph> &graph, int width,
                         int height, int iterations, int x, int y);
 };
